IT DOES NOT WORK YET!!!!!!!!!! Add history and movement handling to the pathfinder
Introduced a new 'History' feature that tracks steps taken by the pathfinding algorithm. This includes the addition of new classes 'Move', 'Step', 'Path', and 'History'. Modified the Pathfinder and Map classes to use and manage this new history. The change will enhance the robustness of the pathfinding process and provide better handling of pathing choices.